requirements.txt
numpy==1.21.2
opencv-python==4.5.3.56
websockets~=10.0
Pillow~=8.3.2